Semi huge logic refactor: 1. Simplified the rendering loop. 2. btrbar should now handle multiple outputs(monitors). 3. It should be able to recover from sudden monitor removals or    additions. 4. It should handle compositor reset (aka switch to TTY and back). 5. Refactored the block handling to share a state between multiple    outputs. 6. Minor code style fixes and addtions.

+++ include/tags.h
@@ -6,7 +6,7 @@
 /*   By: julmajustus <julmajustus@tutanota.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/24 20:22:40 by julmajustus       #+#    #+#             */
-/*   Updated: 2025/07/25 03:44:24 by julmajustus      ###   ########.fr       */
+/*   Updated: 2025/08/09 02:03:29 by julmajustus      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,10 @@
 typedef struct {
 	char		name[MAX_LABEL_LEN];
 	const char	*icon; 
-	int			occupied;
-	int			active;
-	int			urgent;
-	int			x0, x1;
+	uint8_t		occupied;
+	uint8_t		active;
+	uint8_t		urgent;
+	uint32_t	x0, x1;
 } tag_t;
 
 int		init_tags(tag_t *tags);
